refactored read_selected_numbers() #6
