+++ cloud/functions/run_manager/src/manager/bigquery.py
@@ -1,23 +1,25 @@
 import os
 
-import google.cloud.bigquery as bq
+import pandas as pd
+
+from manager.tools import assert_values_in_range
 
 
-def get_selected_numbers(_path: str = 'selected.numbers'):
-    # TODO: GCP read from BigQuery
+def get_selected_numbers():
     """ validates Selected Numbers. Ensures numbers selected are valid and there are no duplicates in Name col."""
-    bq_client = bq.Client()
     select_all_query = f"""
     SELECT *
-    FROM `{os.getenv('PROJECT_ID')}.selected.numbers`
-    """
-    selected_numbers = (
-        bq_client.query(select_all_query) \
-            .result()\
-            .to_dataframe()
-    )
-    bq_client.query()
-    # query bq,
-    # convert to dataframe
-    # perform checks
-    pass
+    FROM `{os.getenv('PROJECT_ID')}.selected.numbers`"""
+    selected_numbers = pd.read_gbq(select_all_query)
+
+    if not selected_numbers['Name'].is_unique:
+        duplicates = selected_numbers['Name'].value_counts()
+        raise ValueError(f'"Name" in Selected numbers needs to be unique. Correct:\n{duplicates[duplicates > 1]}')
+
+    number_cols = [col for col in selected_numbers.columns if col.startswith('Number_')]
+    assert_values_in_range(selected_numbers, start=1, end=50, cols=number_cols)
+
+    star_cols = [col for col in selected_numbers.columns if col.startswith('Lucky_Star_')]
+    assert_values_in_range(selected_numbers, start=1, end=12, cols=star_cols)
+
+    return selected_numbers
