+++ cloud/functions/run_manager/tests/test_tools.py
@@ -4,7 +4,7 @@ from unittest import mock
 import pytest
 
 from manager.tools import (
-    get_last_friday_date, get_folder_name, get_selected_numbers, assert_values_in_range, currency_to_int
+    get_last_friday_date, assert_values_in_range, currency_to_int
 )
 
 
@@ -22,24 +22,6 @@ from manager.tools import (
 )
 def test_get_fri_date_str(today, last_fri, last_fri_str):
     assert get_last_friday_date(today) == (last_fri, last_fri_str)
-
-
-@pytest.mark.parametrize(
-    'a_date, folder_name',
-    [
-        (date(2020, 2, 16), '2020-02-07__2020-03-06'),
-        (date(2020, 3, 23), '2020-03-06__2020-04-03'),
-        (date(2020, 6, 11), '2020-05-29__2020-06-26'),
-        (date(2020, 8, 21), '2020-07-24__2020-08-21'),
-        (date(2020, 9, 17), '2020-08-21__2020-09-18'),
-        (date(2020, 10, 4), '2020-09-18__2020-10-16'),
-        (date(2021, 11, 16), '2021-11-12__2021-12-10'),
-        (date(2021, 12, 5), '2021-11-12__2021-12-10'),
-        (date(2022, 4, 9), '2022-04-01__2022-04-29')
-    ]
-)
-def test_get_folder_name(a_date, folder_name):
-    assert get_folder_name(a_date) == folder_name
 
 
 @pytest.fixture
@@ -85,18 +67,6 @@ def test_assert_values_in_range_raise_key_error(selected_dataframe):
     cols = ['Number 1', 'Number 5']
     with pytest.raises(KeyError, match=r'.* not in date.columns.'):
         assert_values_in_range(selected_dataframe.astype(str), start=1, end=9, cols=cols)
-
-
-def test_get_selected_numbers_raise_value_error(selected_dataframe):
-    with mock.patch('pandas.read_excel', mock.MagicMock(return_value=selected_dataframe)):
-        with pytest.raises(ValueError, match=r'"Name" in Selected numbers needs to be unique\. .*'):
-            get_selected_numbers()
-
-
-def test_get_selected_numbers_valid_run(selected_dataframe):
-    selected_dataframe.iloc[0, 0] = 'Bounty'
-    with mock.patch('pandas.read_excel', mock.MagicMock(return_value=selected_dataframe)):
-        assert get_selected_numbers().equals(selected_dataframe)
 
 
 def test_currency_to_int():
